sturtle/requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
Deprecated==1.2.14
distro==1.8.0
exceptiongroup==1.2.0
fastapi==0.104.1
fief-client==0.18.3
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
idna==3.4
Jinja2==3.1.2
jwcrypto==1.5.1
MarkupSafe==2.1.3
openai==1.3.7
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
#starlette==0.33.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
wrapt==1.16.0
